@@ -3,8 +3,9 @@
 import { useRouter } from "next/navigation";
 import deleteComment from "../actions/deleteComment";
 
-export default async function DeleteButton({ commentId }) {
+export default async function DeleteButton(commentId) {
   const router = useRouter();
+
   async function handleCommentDelete(e) {
     const id = e.target.dataset.id;
     if (id) {
@@ -15,8 +16,9 @@ export default async function DeleteButton({ commentId }) {
 
   return (
     <button
+      role="delete-comment-button"
       className="delete-button"
-      data-id={commentId}
+      data-id={commentId.commentId}
       onClick={(e) => handleCommentDelete(e)}
     >
       ×
