@@ -1,8 +1,8 @@
-"use client"; // Error components must be Client Components
+"use client";
 
 import { useEffect } from "react";
 
-export default function errorComponent({ error, reset }) {
+export default function errorComponent({ error }) {
   useEffect(() => {
     console.error(error);
   }, [error]);
@@ -12,7 +12,7 @@ export default function errorComponent({ error, reset }) {
       <h2>We're Sorry!</h2>
       <p>Something seems to have gone wrong.</p>
       <p>Please click the try again button, or try refreshing the page.</p>
-      <button onClick={() => reset()}>Try again</button>
+      <button onClick={() => window.location.reload()}>Try again</button>
     </div>
   );
 }
