@@ -1,15 +1,11 @@
-import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
-import { cookies } from "next/headers";
+import serverClient from "../../actions/serverClient";
 import PostComments from "../../actions/PostComments";
 import CommentForm from "../../CommentForm";
 import getPost from "@/app/actions/getPost";
 import Post from "@/app/components/Post";
 
 export default async function Page({ params: { slug } }) {
-  const supabase = createServerComponentClient({ cookies });
-  const {
-    data: { session },
-  } = await supabase.auth.getSession();
+  const { session } = await serverClient();
 
   const post = await getPost(slug);
 
@@ -23,7 +19,7 @@ export default async function Page({ params: { slug } }) {
     let liked = false;
     if (!!post.likes && !!post.likes[0]) {
       for (let like of post.likes) {
-        if (like.user_id === session.user.id) {
+        if (session && like.user_id === session.user.id) {
           liked = true;
           break;
         }
